Facilitate the SFO SSO endpoint with ADFS request support
The ADFS Helper service was created to determine if the Http request
that was posted to the server contained a valid Adfs request. If this
is the case. The Helper transforms the Adfs request to an AuthnRequest.

Invalid ADFS requests are served of as invalid requests and follow the
normal error resolving path.