Add a select second factor screen
When the user has more than one viable second factor tokens, a form is
shown to the user. Enabling them to select their favourite token to give
a second factor with.

For now this form is kept really simple. The user can select the
preferred second factor in a radio group and can the submit the form.

The form can also be cancelled, redirecting the user to the
user-cancelled-the-request endpoint.

Translations have been added for the form.