Base matching institutions on users SHO
Searching the institutions of the authenticating user is now based on
the vetted tokens of said user. A token is vetted by an institution,
this information was used to determine the LOA for the authenticating
user.

This change includes the users institution (based on SHO from IdP) to
check against. Ensuring the SHO retrieved from the IdP is also included
in the list of institutions the user is enrolled to.

Tests where added to ensure good coverage of this business logic is
ensured. Thanks to @pmeulen for providing me with a test strategy.

See pivotal for detailed background information.
https://www.pivotaltracker.com/story/show/153594101