Handle impossible SF action more graceful
When the authenticating user cannot provide a sufficient token either
because it was vetted at the wrong institution or the user did not
have any tokens at all. The controller wil now forward these occurrences
to the sendLoaCannotBeGiven action of the gateway controller. This
prevents the display of a generic error page. But shows the end user
with the message: 'You are not authorised to log in.'

To achieve this, LoaCannotBeGivenExceptions are thrown instead of the
more severe runtime exceptions. The domain exception is caught in the SF
controller and forwards the authenticating user to the correct endpoint.

Test expectancies have been updated to the correct exception types.

Cherry picked from: 82ae29b