Update translation bundle to new version
The previous version of the JMS translation bundle (1.1) did not
support PHP 5.6 which we now require because of the recent changes in
the stepup SAML bundle. This commit updates the translation bundle to
version 1.3, fixing the broken translation:extract command.

The XLIFF files are regenerated because the data in these files are
slightly changed in version 1.3.