Refactor resolveHighestRequiredLoa method
The workflow has been revised based on the diagram provided by
@pmeulen. The method has been relieved of several responsibilities and
dependencies:

 * User SHO determination based on vetted tokens is removed
 * $identityNameId is no longer required
 * A ServiceProvider instance is no longer required
 * NULL is no longer a possible return type. Either a Loa is return or
   an exception is thrown.

The StepUpAuthenticationServiceTest was also refactored with the intent
to reuse as many tests as possible. Some tests need to be validated with
the product owner as some test outcomes no longer match the previous
behaviour of the method. This will be addressed in an upcoming commit.

More context can be found in Pivotal:
https://www.pivotaltracker.com/story/show/153594101