Update the resolveHighestRequiredLoa method call
The signature of resolveHighestRequiredLoa is changed slightly. Some
data retrieval responsibilities where moved to the controller (reading
of user schac home organization based on vetted second factors). And
other parameters have been depricated (nameID).

The method now either returns a Loa instance or throws a
LoaCannotBeGivenException exception (or a more specific derivative of
that type). This resulted in the removal of some code from the
controller.

In the WAYG action (what token do you want to use, when multiple tokens
can be registrated by step up users) was also updated to conform to the
changes to the resolveHighestRequiredLoa method.

(cherry picked from commit f47f726)