Assert SHO values are valid
By adding an assertion to the StepUpAuthenticationService, we gained
the possibility to validate the SHO value. A valid SHO value is a
lower cased (or empty) string. When an invalid SHO is tested, an
InvalidStepupShoFormatException is raised. Tests where added to validate
the correct working of the method.

In this commit the previous SHO normalization was undone.