Update LoA determination test assertions
The assertions have been updated according to the expected test results
described in the flow diagram and the original test matrix provided by
@pmeulen.

Tests on case insensitivity have been dropped in the test matrix as this
is not part of the method under test. This is covered elsewhere.

(cherry picked from commit 5bd9677)