Remove the rogue Second Factor migration
The Second Factor schema was not completely in sync with the Middleware
version of the Second Factor entity. Also the migration to sync this
difference was added to the Gateway migrations. This should not have
been as the migrations for the Second Factor projection is managed by
Middleware.

To prevent this from happening in the future I've added a warning for
future self.