Allow mixed-case schacHomeOrganization values
Commit 94080f2 introduced behaviour to the gateway were an error was
thrown on mixed-case SHO values. This behaviour is reverted in this
commit: mixed-case SHO is allowed and compared in an case-insensitive
manner.

While the case of SHO values of the SP/IDP no longer matters for
StepUp, it is important for the SP configuration (SP-configured LoAs)
to always have lower-cased SHOs.

See https://www.pivotaltracker.com/story/show/154845486