Clear previous authentication state when proxying GSSP requests
When stepup acts as a proxy for a GSSP authentication or registration
action, state from previous authentication requests was not properly
cleared. This caused two bugs:

1. Requested Subject NameID "%s" and Response NameID "%s" do not match

A GSSP registration request does not contain a NameID. Instead of
handling the request without NameID, the NameID of the previous
authentication request was used resulting in above error.

2. Unknown Generic SAML Stepup Provider requested "%s", known providers...

When handling a GSSP registration request, the flag
"secondFactorVerificationRequested" set by the previous authentication
action was not reset, causing above error when handling an error in
the GSSP request (for example, when bug 1. was triggered).

This commit explicitly clears the state when the single-sign-on action
is started.

Other parts of Stepup (regular SSO, SFO) are not vulnerable to this
bug because they always set the complete state in one swoop.

See: https://www.pivotaltracker.com/story/show/155774630