Clear previous authentication state on normal SSO and SFO requests
The problem where states were mixed up when handling GSSP requests was
resolved by explicitly clearing the state on every new SSO
request. This was resolved in PR #138.

The problem does not exist on normal SSO/SFO requests, but the session
state can in theory still be inconsistent in those cases. It's also
prone to errors on future changes.

This commit streamlines the way state is cleared: when state is stored
when a GSSP SSO, SSO or SFO initiated, the state is first always
cleared to prevent the possibility of previous state being available
in the new SSO request.

See: https://www.pivotaltracker.com/story/show/155774630