Use ACS URL from AuthnRequest instead of configuration
Instead of using the first configured ACS URL, use the one from the SP
and verify it's allowed in the list of configured URLs. If the SP URL
is not allowed, log a warning and fall back to the first ACS URL in
the configuration.

ADFS is handled slighly different. In the previous situation stepup
always used the ACS URL in the AuthnRequest for ADFS, without
verification. The ADFS behaviour is now similar to the other
use-cases, so if a SP using ADFS is not properly configured they might
stop working. The validation scheme for ADFS is different from the
regular scheme: the ACS URL must start with the configured ACS
location, but may contain extra characters (like query parameters).

Working on this feature brought a slighly improved understanding of
how the SamlProxyController works, and the docblocks and diagrams in
GatewayState.md have been updated accordingly.

See: https://www.pivotaltracker.com/story/show/155288768