Restore samlstepupproviders.yml
The composer post-install hooks require valid configuration, and since
samlstepupproviders.yml is not a parameters file (managed by
incenteev-parameters) it is not automatically created on install.

Better solutions would be:

 - copy .yml.dist -> .yml in a post-install hook
 - or refactor the file so they are true parameters

For now, we simply revert to the old behaviour. Note: SS and RA also
have a samlstepupproviders.yml in git.

Reverts part of a5389ed.