Document main Stepup flows
The docs/GatewayState.md file contains diagrams showing how various
parts of Stepup interact with outside actors and other parts of
Stepup. This fills a gap between more abstract diagrams in the SURF
wiki and Stepup-Deploy wiki, and the actual code. The diagrams show
the various controllers and controller actions and what information is
read from and stored to the user session at a given point in the flow.

The most important controller actions are now annotated with a
description of how they are used.

See: https://www.pivotaltracker.com/story/show/155969747