Improve GSSP authentication error handling
This is a spin-off from PR #155. It is hard to implement the changes
requested there because the current implementation is incomplete /
buggy. So fixing it here first so we can properly test #155.

When the user cancels a GSSP request, or any other error on the GSSP
side, the user landed on the gateway on a generic error page. The
implementation was intended to show an error page with a "back to SP"
button.

This commit fixes the issue by always returning the user to the
originating SP, without intermediate error page on the gateway.

What the originating SP is in consumeAssertionAction() depends on the
situation:

 - in case of registration, the user lands in ssoAction directly from
   the SP (selfservice), and the SP entity ID is remembered in the
   GSSP state

 - in case of verification, the user is redirected internally to
   sendSecondFactorVerificationAuthnRequestAction() where the
   originating SP is unknown. In consumeAssertionAction() we check if
   the request is for verification, and if so, get the response
   context from the GatewayController where the initial AuthnRequest
   landed, and read the originating SP entity ID from there