Use assertion NotOnOrAfter for SubjectConfirmation validity
The SAML spec states the NotOnOrAfter of a subject confirmation should
not exceed the value of the assertions NotOnOrAfter. Gateway now uses
the assertion NotOnOrAfter value of SubjectConfirmation, instead of a
timestamp 8 hours in the future. The NotOnOrAfter is now always 5
minutes after generating the response.

See https://www.pivotaltracker.com/story/show/157880479