Load contents of .env file when available
The concept is to load the .env file from the application root. If it
cannot be found there, try to fall back on a Stepup globally set .env.
If this does not exist either, just commit the default values to PHPs
environment.

The previously installed vlucas dotenv component is used to load and
parse existing .env files. In a perfect world we would utilize Symfony
DotEnv. For now we're stuck on version 2.7 and have to make due with
this component.