Remodel the StepUpAuthenticationService
Some of the application logic was moved to the API, this to be able to
mock the Api service. In essence the YubikeyOtpVerificationResult was
moved to the API, becoming the return type for said service.

The verification logic that was performed in the Gateway bundle was
moved to the ApiBundle.

This was the most elegant way to then provide a test double. An
alternative was to also provide a test double for the
StepUpAuthenticationService, which is a very godlike service which would
result in losing too much important business logic for the E2E tests.

See: https://www.pivotaltracker.com/story/show/156570267