Split verification logic to support all usages
In the previous work on the YubikeyService, in an attempt to make the
service replaceable by a test double, the
YubikeyController::verifyAction was overlooked. This method is not able
to provide the token public id (identifier) as this token is yet to be
registered.

So the logic was once again separated into seperate methods that verify otp and the public key.