Add subcode and message to AuthNFailed GW response
The GW proxy controller now also returns the sub code and error message
when one of the GSSP returned a failed response.

The exception message that was thrown while processing the SAMLResponse
is set as the message yielding something like:

"Could not process received Response, error: "Cannot process response,
preconditions not met: "Responder/AuthnFailed User cancelled the
request""

This is far from perfect, preferably we'd simply set the original
"User cancelled the request" message, this would entail quite some
rework to the gateway.