Use correct ResponseContext service identifier
Between regular SSO and SFO authentication, different response context
objects are used. It is very important that the correct one is used
(retrieved from container) during the ACS step. This was not the case
when dealing with SFO responses. In this commit, that is solved.

The problem is described in greater detail in the bug report:
https://www.pivotaltracker.com/story/show/168746898