Require authenticationMode on every action
Each SFController action now requires the authentication mode to be set. With this knowledge they are able to load the correct response context from the session state handler.

This is not the prettiest of solutions. Other considerations where to handle this in an event handler, triggered right before going into the action. A controller middleware so to speak. This however also adds to the (unseen) complexity of the controller. This solution is somewhat more transparent and therefore, hopefully more understandable.