Run behat tests on Github Actions
Gatway now runs in a dockerized environment with a database server, nginx and php-fpm container. Also a selenium container with headless Chrome has been prepared for the multiple in flight sessions tests.

A ci folder was created with CI specific config files

A .github folder was added containing the yaml file that specifies the test integration.