Update default environments
By default the app.php, when the APP_ENV is not in the environment, will
use prod as its default environment. The app_dev.php.dist file was
updated in accordance to that of the app.php, but with 'dev' as its
default environment.

Later the app.php was reverted to a state before release-17.1.

See task 1 on: https://www.pivotaltracker.com/story/show/172199429