Stop using Mopa Bootstrap
The dependency on Mopa was moved to NPM packages. Previously we utilized
the MopaBootstrap bundle and Twb Bootstrap to pull bootstrap and
fontawesome. We also used some of the mopa form theming.

This was replaced by self written/copied counterparts.

The Symfony form theming, bootstrap 4 style was also utilized to almost
get a pixel perfrect replacemt without Mopa Bootstrap.