Remove useless step definition
The step definition might look allright in the context of the scenario,
but does not actually do anything in the context of the gateway. In a
broader Stepup sense, it would be required for the user to exist at the
IdP, but gateway is not aware of this. It only knows about token
ownership.