Abstract the SmsService
In order to use both MessageBird and Spryng, the existing MessageBird
tight SMS service needs some decoupling.

This commit ensures the SMS service and the message send result are
abstracted. This way the application can safely use other adaptations of
the SMS implementation of the Gatway. Effectively decoupling the
application from the infrastructure.