Use internal-collabPersonId in state handler
When receiving the assertion from the IdP, we store some of the data
from the assertion in the state handler for later use. Evaluation of the
nameid was not performed correctly. The internal-collabPersonID was not
evaluated for containing the unspecified name id of the identity. But
would get the subject nameid which might be configured to release a
persistent or even transient identifier.

For a detailed bug report and some discussion on this problem, please
consult the ticket linked below.

See https://www.pivotaltracker.com/story/show/182274353