Make NameId handling in state handler more robust
Having the name id on the proxy state handler is a requirement for GW to
perform correctly. The application now deals with this by returning a
null value when the NameID is not present in state. This works out fine,
and results in a pretty good error down the line. But having an error earlier in
the process is better in my opinion.

The previous commit already started throwing an exception on the write
side of the NameId. This commit further hardens the read side. By
checking for a valid value in state, and being more verbose in the
return type declaration department.

Filed under boyscout rule