Configure exceptions to Allow|Store SSO on 2FA
Don't store the SSO cookie when a RA vetting procedure is handled by
Gateway. Or any of the other SP's that have this explicitly configured
in the middleware config.

Don't try to give SSO on 2FA when one of the SPs which are configured to
not evaluate the SSO cookie. This is also configured in the MW config.

See: https://www.pivotaltracker.com/story/show/183511012
See: https://github.com/OpenConext/Stepup-Middleware/pull/391