Implement loa resolution using identity-vetted flag
Instead of having the vetting type like we used to, the SecondFactor
projection now carries a boolean value, indicating if the SF was vetted
using a traditional RA desk vetting method. If this is not the case, we
can assume a self asserted vetting method was used.

The can satisfy and getloa level methods have been adjusted to map the
flag to a faux vetting type that corresponds to the identity-vetted
value of the token.

See: https://www.pivotaltracker.com/story/show/184191063