Set correct Issuer on SAML Failed Response
When the GSSP authentication/registraion failed, the incorrect issuer
was returned to the SP. That was rectified in this commit. The issuer is
loaded from the correct context (cancelled SFO auth would return a SSO
issuer where the SFO issuer should be used instead).

In addition the message that is posted back is also simplified. Only the
Message provided by the remote IdP is used in this case.

See: https://www.pivotaltracker.com/story/show/179591654