Set response context service id on state handler
When starting a gssp registration, the response context service id was
not set on the state handler. Causing the new improved error response
redirection to fail. As the state handler would not be able to load the
correct response context service to get the Issuer entity id from.
Causing the response to end up in a 500 error, and not a 301 back to the
issuer.