Create two SSO 2FA behat tests
In order to get them to work, a host of changes where required. Most
notable changes required was that the SSO authentications did not yet
utilize the mock gateway remote IdP setup. It visited the SSP mock idp
that is not available in this container. Trololo

Next up was to configure the certificates correctly.

Finally the new sso-on-2fa.feature itself could be constructed.