Reinstate the Selenium behat tests
Test one and two have been fixed. The Yubikey and SMS tests could be
enabled in the new (SSP based) test environment. The Tiqr variant could
not yet be enabled as Tiqr is not functioning in that environment yet.

During enabling the SMS test, I found out that the SMS Bypass fix was
not taken into account in the existing sfo test. That has been updated.