Validate the cookie value authentication time
The cookie lifetime can be extended by the user. But the explicit
lifetime of the cookie was not yet validated by the cookie service.
Meaning you could extend the SSO lifetime by keeping your cookie alive.

As of this change, the authentication lifetime is checked against the
current timestamp togegher with the cookie expiration time.

See: https://www.pivotaltracker.com/story/show/183402734
 -> the SSO cookie's timestamp is within the configured SSO lifetime