Repair SFO SSO on 2FA handling
The state handler lost the SF ID to early in the process and that caused
the saving of the sso cookie to fail. This was introduced when I added
the check on the SF token still existed in:

https://github.com/OpenConext/Stepup-Gateway/pull/294

The new way of handling this, is to remove the SF id at the last
possible moment.