Recreate SSO cookie after real authentication
When a real authentication takes place, the sso cookie must be
rewritten. This ensures it stays alive. This also makes it possible that
the LoA stored in the SSO cookie is downgraded to a lower level if the
new authentication was performed with a token with a lower level of
assurance.

See: https://www.pivotaltracker.com/story/show/186011497