Ensure consume_assertion.html.twig is rendered once
Previously we would always render the regular consume_assertion.html.twig
template, and additionally render the adfs variant when dealing with an
ADFS response. That caused the JS from not ending up in the template, as
Twig concluded that the (reused) JS was already on the page (as it was
just rendered).

This ensures the render method is called only once. Circumventing the
problem. But not realy solving the issue at hand. Which is that the
controller logic is getting growingly more complex and harder to
understand.

For that problem, I'm opening a technical debt issue in Pivotal.

https://github.com/OpenConext/Stepup-Gateway/issues/312