Prevent state issues by actually clearing the state
As we no longer use the namespaced attribute bag, we now need to
manually clear the gssp prefixed values from the state handler.

This clear method was not implemented correctly, leaving all values in
state. Creating some state confusion issues when executing sequential
authentication or registrations with a GSSP token of a certain type.

By clearing the state, those issues are resolved