Run local acceptance tests in `test` env
They used to run on the smoketest env. And that was fine, until the
suite wide acceptance tests started using that env. Leaving us with some
conflicting setup. Major changes are: the smoketest tests use the
devconf docker compose setup to run gateway. So many certificates and
hostnames are different from the one we use on CI.

So now:
- CI runs the ci/docker docker compose. And those tests run in
`test` env
- The devconf behat tests run in `smoketest`.