Remove session bag from Gssp StateHandler
Using an additional attribute bag for gssp state is a good idea. It used
to compartmentalize the GSSP state data from the other GW state.

Now that registration of session bags has been made very difficult,
almost discouragingly so. I decided to store the session state in the
same (default) bag. Setting, clearing and getting is done via an unique
prefix. So state of other authentication can not be mingled.