Improve service locator retrieval logic
The get method on the ContainerController left some to the imagination
when it comes to logging. This commit adds logging and makes transparant
which service was being loaded. This is usefull in case when the service
could not be loaded.

The other part is that the SF controller when the ResponseContext is
loaded. The unavailability of the StateHandlers response context id is
now checked.