Improve setup instructions & fix running the functional tests
Prior to this change, running the functional behat tests would fail after a clear-cache.
This happened because clear-cache removes the cache dir, and recreates it as root (because cache-clear is run as root).

This causes the web php runner to not be able to write the cache, resulting in errors. This causes the functional tests to fail.

Also scoutfix the phpcbf not fixing anything by default. It should match
phpcs.