Upgrade frontend build
Prior to this change, eslint used the `airbnb` config. This is a config for react projects. Stepup-Gateway does not use react, resulting in errors, because eslint required things like jsx packages.

This change uses `airbnb-base` and adds the required ts support.

---

Resolve Sass compilation warnings by running the migrate tool

--

Replaced `eslint-loader` with `eslint-webpack-plugin` as eslint-loader is deprecated, and suggests eslint-webpack-plugin as a replacement.