Fix some sass deprecation warnings.
Silence the rest, as it originates from bootstrap 5, which we do not have control over.

Keep an @import for bootstrap in the application.scss as bootstrap is
not compatible with newer sass.

Also switch sass-loader to use the modern Dart Sass JavaScript API instead of the legacy one.

This resolves Module Warning (from ./node_modules/sass-loader/dist/cjs.js):
              Deprecation The legacy JS API is deprecated and will be removed in Dart Sass 2.0.0.

              More info: https://sass-lang.com/d/legacy-js-api