Legacy browser support
Prior to this change, the js + css would target modern browsers.

This change adds autoprefixer to output css for older browsers and babel
for the js.

Realistically, ie6 support is not possible. This change ensures the
modern features of css and js are kept back as much as possible.
As the templates function without javascript, really old browsers should
click perform the redirect automatically. Apart from autosubmit, no js
is used.