Remove deploy connection from doctrine dbal
Prior to this change, doctrine still had a deploy connection, just because the gateway_test schema needed to be created.
All other write interaction with the gateway_test was performed via PDO.
Also, the gateway_test is currently based on the current entity mapping. But because the middleware generated schema is not available in the pipeline, this cannot be changed.

This change removes the deploy connection from doctrine, and creates the schema by using PDO.

It also refactors the test db connection that does require write access
to use the same PDO connection as the other test db interaction, instead
of relying on the framework dbal connection. This makes it possible to
remove the deploy connection from doctrine.

Resolves https://github.com/OpenConext/Stepup-Gateway/issues/402